requirements.txt
boto3
duckdb
requests
python-dotenv
jupyter
pandas
numpy